requirements.txt
openai>=1.0.0
pinecone-client>=3.0.0
python-dotenv>=0.19.0
numpy>=1.19.0
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
pydantic>=2.0.0
